add action test to pagination links
