+++ active_model_serializers.gemspec
@@ -24,4 +24,6 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency "bundler", "~> 1.6"
   spec.add_development_dependency "timecop", ">= 0.7"
   spec.add_development_dependency "rake"
+  spec.add_development_dependency "kaminari"
+  spec.add_development_dependency "will_paginate"
 end
